requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.8.0
django-filter==2.4.0
django-restframework==0.0.1
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
